requirements.txt
seaborn
scikit-learn
matplotlib
pandas
flask
tensorflow
pillow
numpy
fpdf
gdown
reportlab

